feat(aci): add uptime details timeline
