+++ static/app/views/alerts/rules/uptime/detailsSidebar.tsx
@@ -5,20 +5,16 @@ import {SectionHeading} from 'sentry/components/charts/styles';
 import {CodeSnippet} from 'sentry/components/codeSnippet';
 import {ActorAvatar} from 'sentry/components/core/avatar/actorAvatar';
 import {Grid} from 'sentry/components/core/layout';
-import {ExternalLink} from 'sentry/components/core/link';
 import {Text} from 'sentry/components/core/text';
 import {KeyValueTable, KeyValueTableRow} from 'sentry/components/keyValueTable';
 import Placeholder from 'sentry/components/placeholder';
-import QuestionTooltip from 'sentry/components/questionTooltip';
-import {t, tct, tn} from 'sentry/locale';
-import {space} from 'sentry/styles/space';
+import {t, tn} from 'sentry/locale';
 import type {UptimeDetector} from 'sentry/types/workflowEngine/detectors';
 import getDuration from 'sentry/utils/duration/getDuration';
-import {CheckIndicator} from 'sentry/views/alerts/rules/uptime/checkIndicator';
-import {CheckStatus, type UptimeSummary} from 'sentry/views/alerts/rules/uptime/types';
+import {DetailsTimelineLegend} from 'sentry/views/alerts/rules/uptime/detailsTimelineLegend';
+import {type UptimeSummary} from 'sentry/views/alerts/rules/uptime/types';
 import {UptimeDuration} from 'sentry/views/insights/uptime/components/duration';
 import {UptimePercent} from 'sentry/views/insights/uptime/components/percent';
-import {statusToText} from 'sentry/views/insights/uptime/timelineConfig';
 
 interface UptimeDetailsSidebarProps {
   showMissedLegend: boolean;
@@ -43,85 +39,10 @@ export function UptimeDetailsSidebar({
         columns={summary && summary.avgDurationUs !== null ? '2fr 1fr 1fr' : '1fr 1fr'}
         gap="2xl"
       >
-        <div>
+        <UptimeContainer>
           <SectionHeading>{t('Legend')}</SectionHeading>
-          <CheckLegend>
-            <CheckLegendItem>
-              <CheckIndicator status={CheckStatus.SUCCESS} />
-              <LegendText>
-                {statusToText[CheckStatus.SUCCESS]}
-                <QuestionTooltip
-                  isHoverable
-                  size="sm"
-                  title={tct(
-                    'A check status is considered uptime when it meets the uptime check criteria. [link:Learn more].',
-                    {
-                      link: (
-                        <ExternalLink href="https://docs.sentry.io/product/alerts/uptime-monitoring/#uptime-check-criteria" />
-                      ),
-                    }
-                  )}
-                />
-              </LegendText>
-            </CheckLegendItem>
-            <CheckLegendItem>
-              <CheckIndicator status={CheckStatus.FAILURE} />
-              <LegendText>
-                {statusToText[CheckStatus.FAILURE]}
-                <QuestionTooltip
-                  isHoverable
-                  size="sm"
-                  title={tct(
-                    'A check status is considered as a failure when a check fails but hasn’t recorded three consecutive failures needed for Downtime. [link:Learn more].',
-                    {
-                      link: (
-                        <ExternalLink href="https://docs.sentry.io/product/alerts/uptime-monitoring/#uptime-check-failures" />
-                      ),
-                    }
-                  )}
-                />
-              </LegendText>
-            </CheckLegendItem>
-            <CheckLegendItem>
-              <CheckIndicator status={CheckStatus.FAILURE_INCIDENT} />
-              <LegendText>
-                {statusToText[CheckStatus.FAILURE_INCIDENT]}
-                <QuestionTooltip
-                  isHoverable
-                  size="sm"
-                  title={tct(
-                    'A check status is considered downtime when it fails 3 consecutive times, meeting the Downtime threshold. [link:Learn more].',
-                    {
-                      link: (
-                        <ExternalLink href="https://docs.sentry.io/product/alerts/uptime-monitoring/#uptime-check-failures" />
-                      ),
-                    }
-                  )}
-                />
-              </LegendText>
-            </CheckLegendItem>
-            {showMissedLegend && (
-              <CheckLegendItem>
-                <CheckIndicator status={CheckStatus.MISSED_WINDOW} />
-                <LegendText>
-                  {statusToText[CheckStatus.MISSED_WINDOW]}
-                  <QuestionTooltip
-                    isHoverable
-                    size="sm"
-                    title={tct(
-                      'A check status is unknown when Sentry is unable to execute an uptime check at the scheduled time. [link:Learn more].',
-                      {
-                        link: (
-                          <ExternalLink href="https://docs.sentry.io/product/alerts/uptime-monitoring/#uptime-check-failures" />
-                        ),
-                      }
-                    )}
-                  />
-                </LegendText>
-              </CheckLegendItem>
-            )}
-          </CheckLegend>
-        </div>
+          <DetailsTimelineLegend showMissedLegend={showMissedLegend} />
+        </UptimeContainer>
         {summary?.avgDurationUs !== null && (
           <div>
             <SectionHeading>{t('Duration')}</SectionHeading>
@@ -200,29 +121,8 @@ export function UptimeDetailsSidebar({
   );
 }
 
-const CheckLegend = styled('ul')`
-  display: grid;
-  grid-template-columns: max-content 1fr;
-  margin-bottom: ${space(2)};
-  padding: 0;
-  gap: ${space(1)};
-`;
-
-const CheckLegendItem = styled('li')`
-  display: grid;
-  grid-template-columns: subgrid;
-  align-items: center;
-  grid-column: 1 / -1;
-`;
-
-const LegendText = styled('div')`
-  display: flex;
-  gap: ${space(1)};
-  align-items: center;
-`;
-
 const MonitorUrlContainer = styled('div')`
-  margin-bottom: ${space(2)};
+  margin-bottom: ${p => p.theme.space.xl};
 
   h4 {
     margin-top: 0;
@@ -230,5 +130,5 @@ const MonitorUrlContainer = styled('div')`
 `;
 
 const UptimeContainer = styled('div')`
-  margin-bottom: ${space(2)};
+  margin-bottom: ${p => p.theme.space.xl};
 `;
