+++ static/app/views/detectors/components/details/uptime/index.tsx
@@ -1,3 +1,5 @@
+import {useCallback, useState} from 'react';
+
 import {KeyValueTableRow} from 'sentry/components/keyValueTable';
 import {DatePageFilter} from 'sentry/components/organizations/datePageFilter';
 import DetailLayout from 'sentry/components/workflowEngine/layout/detail';
@@ -6,6 +8,12 @@ import {t} from 'sentry/locale';
 import type {Project} from 'sentry/types/project';
 import type {UptimeDetector} from 'sentry/types/workflowEngine/detectors';
 import getDuration from 'sentry/utils/duration/getDuration';
+import {DetailsTimeline} from 'sentry/views/alerts/rules/uptime/detailsTimeline';
+import {DetailsTimelineLegend} from 'sentry/views/alerts/rules/uptime/detailsTimelineLegend';
+import {
+  CheckStatus,
+  type CheckStatusBucket,
+} from 'sentry/views/alerts/rules/uptime/types';
 import {UptimeChecksTable} from 'sentry/views/alerts/rules/uptime/uptimeChecksTable';
 import {DetectorDetailsAssignee} from 'sentry/views/detectors/components/details/common/assignee';
 import {DetectorDetailsAutomations} from 'sentry/views/detectors/components/details/common/automations';
@@ -20,12 +28,24 @@ type UptimeDetectorDetailsProps = {
 export function UptimeDetectorDetails({detector, project}: UptimeDetectorDetailsProps) {
   const dataSource = detector.dataSources[0];
 
+  // Only display the missed window legend when there are visible missed window
+  // check-ins in the timeline
+  const [showMissedLegend, setShowMissedLegend] = useState(false);
+
+  const checkHasUnknown = useCallback((stats: CheckStatusBucket[]) => {
+    const hasUnknown = stats.some(bucket =>
+      Boolean(bucket[1][CheckStatus.MISSED_WINDOW])
+    );
+    setShowMissedLegend(hasUnknown);
+  }, []);
+
   return (
     <DetailLayout>
       <DetectorDetailsHeader detector={detector} project={project} />
       <DetailLayout.Body>
         <DetailLayout.Main>
           <DatePageFilter />
+          <DetailsTimeline uptimeDetector={detector} onStatsLoaded={checkHasUnknown} />
           <Section title={t('Recent Check-Ins')}>
             <div>
               <UptimeChecksTable
@@ -41,6 +61,9 @@ export function UptimeDetectorDetails({detector, project}: UptimeDetectorDetails
           <Section title={t('Detect')}>{t('Three consecutive failed checks.')}</Section>
           <Section title={t('Resolve')}>
             {t('Three consecutive successful checks.')}
+          </Section>
+          <Section title={t('Legend')}>
+            <DetailsTimelineLegend showMissedLegend={showMissedLegend} />
           </Section>
           <DetectorDetailsAssignee owner={detector.owner} />
           <DetectorExtraDetails>
